Update reports automatically on Agora
Summary:
Rhyne-Wyse is a bot based on pywikibot to update pages on the wiki
for automated reports.

It can use the code from nasqueron-reports tool, the same `run-report` uses,
to generate a report for ressources reachable from a standard server.

For resources less easily accessible, it can also fetch a report
already generated by `run-report` on a remote URL, as implemented
in D3674 for AI content from DevCentral databases.

It's not intended to run as a daemon like an IRC bot, but to be run
at regular intervals, for example daily.

Ref T2124

Test Plan: https://agora.nasqueron.org/?diff=1977

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T2124

Differential Revision: https://devcentral.nasqueron.org/D3678