Configure Vault connection with secretsmith
Summary:
The new secretsmith package is a wrapper on the top of hvac to configure
connection and authentication parameters to Vault/OpenBao.

It's also the opportunity to publish our read_secret function,
used everywhere from Sentry custom config to those reports tools.

Ref T2124

Test Plan: Deploy on WindRiver to authenticate to Vault with AppRole

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T2124

Differential Revision: https://devcentral.nasqueron.org/D3697